Limit available regions to those supported by SES
For simplicity, we now create every resource in the same region
supported by SES. This is necessary, for instance, because Lambda
functions created in a different region than the one used for SES cannot
be used as actions in receipt rules. Also, Lambda functions cannot
be created in a different region than the S3 bucket used to provide the
source code of the function.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>